refactor: split oversized files to stay under 210-line limit
Extract token extractors from queryParsing.ts into tokenExtractors.ts,
OR handler helpers from orExpansion.ts into orHelpers.ts, and
query translation helpers from QueryTranslation.tsx into
queryTranslationHelpers.ts.
